feat: calendar, sidebar 폴더 변경
